# docs/openapi.yaml
openapi: 3.0.0
servers:
  - description: CERCLES API
    url: http://localhost:8080
info:
  version: "1.0.0"
  title: CERCLES API Documentation
  description: >-
    Esta API proporciona los endpoints para la aplicación CERCLES.
components:
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Ingrese su token JWT para autenticar las solicitudes"
security:
  - BearerToken: []
...
paths:
...
  /api/cursos/crear:
    post:
      tags:
        - Cursos
      summary: Crear un nuevo curso
      description: >
        Este endpoint permite a un profesor crear un nuevo curso. Se deben de asignar profesores,
        estudiantes y establecer los periodos de evaluación para el curso.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombreAsignatura:
                  type: string
                  description: El nombre de la asignatura del curso.
                añoInicio:
                  type: integer
                  description: El año de inicio del curso.
                cuatrimestre:
                  type: integer
                  description: El cuatrimestre en el que se imparte el curso.
                activo:
                  type: boolean
                  description: Indica si el curso estará activo.
                profesores:
                  type: array
                  items:
                    type: integer
                  description: Lista de IDs de los profesores que serán asignados al curso.
                estudiantes:
                  type: array
                  items:
                    type: integer
                  description: Lista de IDs de los estudiantes que serán asignados al curso.
                periodosEvaluacion:
                  type: array
                  items:
                    type: object
                    properties:
                      fechaInicio:
                        type: string
                        format: date
                        description: Fecha de inicio del periodo de evaluación.
                      fechaFin:
                        type: string
                        format: date
                        description: Fecha de finalización del periodo de evaluación.
      responses:
        "201":
          description: Curso creado exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      description: El identificador único del curso.
        "400":
          description: Solicitud inválida
          content:
            text/plain:
              schema:
                type: string
                example: Datos de solicitud no válidos.
        "403":
          description: Usuario no autenticado o sin permisos
          content:
            text/plain:
              schema:
                type: string
                example: Usuario no autenticado o sin permisos.
        "409":
          description: Conflicto (Curso activo ya existente)
          content:
            text/plain:
              schema:
                type: string
                example: Curso activo ya existente. ¿Desea desactivarlo?
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example: Error al crear el curso.
  /api/cursos/{id}:
    get:
      tags:
        - Cursos
      summary: Obtener detalles de un curso
      description: >
        Este endpoint permite a un profesor obtener detalles completos de un curso específico. 
        Incluye información sobre los estudiantes sin equipo, los equipos con sus miembros y los profesores asociados al curso.
      parameters:
        - name: id
          in: path
          required: true
          description: El identificador único del curso cuyos detalles se desean obtener.
          schema:
            type: integer
      responses:
        "200":
          description: Detalles del curso obtenidos exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: El identificador único del curso.
                  nombreAsignatura:
                    type: string
                    description: El nombre de la asignatura del curso.
                  añoInicio:
                    type: integer
                    description: El año de inicio del curso.
                  cuatrimestre:
                    type: integer
                    description: El cuatrimestre en el que se imparte el curso.
                  activo:
                    type: boolean
                    description: Indica si el curso está activo.
                  estudiantesSinGrupo:
                    type: array
                    items:
                      type: object
                      properties:
                        nombre:
                          type: string
                          description: Nombre del estudiante sin equipo.
                        correo:
                          type: string
                          description: Correo electrónico del estudiante sin equipo.
                  equipos:
                    type: array
                    items:
                      type: object
                      properties:
                        nombre:
                          type: string
                          description: Nombre del equipo.
                        id:
                          type: integer
                          description: Identificador del equipo.
                        evaluadorId:
                          type: integer
                          description: Identificador del evaluador del equipo (si existe).
                        miembros:
                          type: array
                          items:
                            type: string
                            description: Nombre de los estudiantes en el equipo.
                  profesores:
                    type: array
                    items:
                      type: string
                      description: Nombre de los profesores asociados al curso.
        "403":
          description: Acceso denegado
          content:
            text/plain:
              schema:
                type: string
                example: Usuario no autenticado o sin permisos.
        "404":
          description: Curso no encontrado o acceso denegado
          content:
            text/plain:
              schema:
                type: string
                example: Curso no encontrado o no tiene acceso a este curso.
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example: Error al obtener los detalles del curso.
  /api/cursos/{id}/modificar_curso:
    put:
      tags:
        - Cursos
      summary: Modificar un curso existente
      description: >
        Este endpoint permite a un profesor modificar un curso existente. Los cambios incluyen la actualización
        de los datos básicos del curso, la adición o eliminación de estudiantes y profesores, y el manejo de relaciones asociadas.
      parameters:
        - name: id
          in: path
          required: true
          description: El identificador único del curso a modificar.
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombreAsignatura:
                  type: string
                  description: El nombre de la asignatura del curso.
                añoInicio:
                  type: integer
                  description: El año de inicio del curso.
                cuatrimestre:
                  type: integer
                  description: El cuatrimestre en el que se imparte el curso.
                estudiantesAñadir:
                  type: array
                  items:
                    type: object
                    properties:
                      correo:
                        type: string
                        description: El correo del estudiante a añadir.
                      nombre:
                        type: string
                        description: El nombre del estudiante.
                estudiantesBorrar:
                  type: array
                  items:
                    type: object
                    properties:
                      correo:
                        type: string
                        description: El correo del estudiante a borrar.
                profesoresAñadir:
                  type: array
                  items:
                    type: object
                    properties:
                      correo:
                        type: string
                        description: El correo del profesor a añadir.
                      nombre:
                        type: string
                        description: El nombre del profesor.
                profesoresBorrar:
                  type: array
                  items:
                    type: object
                    properties:
                      correo:
                        type: string
                        description: El correo del profesor a borrar.
      responses:
        "200":
          description: Curso modificado exitosamente
          content:
            text/plain:
              schema:
                type: string
                example: Curso modificado exitosamente.
        "400":
          description: Solicitud inválida
          content:
            text/plain:
              schema:
                type: string
                example: Datos de solicitud no válidos.
        "403":
          description: Usuario no autenticado o sin permisos
          content:
            text/plain:
              schema:
                type: string
                example: Usuario no autenticado o sin permisos.
        "404":
          description: Curso no encontrado
          content:
            text/plain:
              schema:
                type: string
                example: Curso no encontrado.
        "409":
          description: Conflicto al actualizar datos del curso
          content:
            text/plain:
              schema:
                type: string
                example: Ya existe un curso activo con el mismo nombre, año y cuatrimestre.
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example: Error al modificar el curso.
  /api/cursos/{id}/borrar:
    delete:
      tags:
        - Cursos
      summary: Borrar un curso existente
      description: >
        Este endpoint permite a un profesor borrar un curso específico identificado por su ID.
        Solo un usuario con rol de profesor puede realizar esta acción.
      parameters:
        - name: id
          in: path
          required: true
          description: El identificador único del curso a borrar.
          schema:
            type: integer
      responses:
        "200":
          description: Curso borrado exitosamente
          content:
            text/plain:
              schema:
                type: string
                example: Curso borrado exitosamente.
        "403":
          description: Usuario no autenticado o sin permisos
          content:
            text/plain:
              schema:
                type: string
                example: Usuario no autenticado.
        "404":
          description: Curso no encontrado
          content:
            text/plain:
              schema:
                type: string
                example: Curso no encontrado.
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example: Error al borrar el curso.
  /api/cursos/verificarCursoExistente:
    post:
      tags:
        - Cursos
      summary: Verificar si un curso activo ya existe
      description: >
        Este endpoint permite a un profesor verificar si un curso con los mismos datos (nombre, año, cuatrimestre) ya existe como curso activo.
        Devuelve un mensaje indicando si hay un conflicto o si no hay ningún curso activo similar.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombreAsignatura:
                  type: string
                  description: El nombre de la asignatura del curso.
                añoInicio:
                  type: integer
                  description: El año de inicio del curso.
                cuatrimestre:
                  type: integer
                  description: El cuatrimestre en el que se imparte el curso.
      responses:
        "200":
          description: No hay conflicto, el curso no existe como activo
          content:
            text/plain:
              schema:
                type: string
                example: No hay conflicto
        "409":
          description: Curso activo ya existente
          content:
            text/plain:
              schema:
                type: string
                example: Curso activo ya existente. ¿Desea desactivarlo?
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example: Error al verificar el curso existente.
  /api/cursos:
    get:
      tags:
        - Cursos
      summary: Obtener cursos asociados a un profesor
      description: >
        Este endpoint permite a un profesor autenticado obtener una lista de los cursos que está gestionando. 
        Devuelve detalles de cada curso, como su nombre, año, cuatrimestre, estado activo y el número de estudiantes inscritos.
      responses:
        "200":
          description: Lista de cursos obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      description: El identificador único del curso.
                    nombreAsignatura:
                      type: string
                      description: El nombre de la asignatura del curso.
                    añoInicio:
                      type: integer
                      description: El año de inicio del curso.
                    cuatrimestre:
                      type: integer
                      description: El cuatrimestre en el que se imparte el curso.
                    activo:
                      type: boolean
                      description: Indica si el curso está activo.
                    numeroEstudiantes:
                      type: integer
                      description: El número de estudiantes inscritos en el curso.
        "403":
          description: Usuario no autenticado o sin permisos
          content:
            text/plain:
              schema:
                type: string
                example: Usuario no autenticado.
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example: Error al obtener los cursos.
  /api/cursos/{cursoId}/equipos:
    get:
      tags:
        - Cursos
      summary: Obtener equipos asociados a un curso
      description: >
        Este endpoint permite obtener una lista de los equipos pertenecientes a un curso específico. 
        Puede ser accedido tanto por estudiantes como por profesores autenticados.
      parameters:
        - name: cursoId
          in: path
          required: true
          description: El identificador único del curso cuyos equipos se quieren consultar.
          schema:
            type: integer
      responses:
        "200":
          description: Lista de equipos obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      description: El identificador único del equipo.
                    nombre:
                      type: string
                      description: El nombre del equipo.
                    cursoId:
                      type: integer
                      description: El identificador del curso al que pertenece el equipo.
                    evaluadorId:
                      type: integer
                      description: El identificador del profesor evaluador del equipo.
                    cursoNombre:
                      type: string
                      description: El nombre del curso al que pertenece el equipo.
                    activo:
                      type: boolean
                      description: Indica si el curso está activo.
        "400":
          description: Solicitud inválida
          content:
            text/plain:
              schema:
                type: string
                example: El curso especificado no existe.
        "403":
          description: Usuario no autenticado o sin permisos
          content:
            text/plain:
              schema:
                type: string
                example: Usuario no autenticado.
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example: Error al obtener los equipos.
...
  /api/cursos/{id}/profesores:
    get:
      tags:
        - Cursos
      summary: Obtener profesores de un curso
      description: >
        Este endpoint permite obtener una lista de los profesores asignados a un curso específico.
        Solo los estudiantes autenticados que pertenecen al curso pueden acceder a esta información.
      parameters:
        - name: id
          in: path
          required: true
          description: El identificador único del curso cuyos profesores se quieren consultar.
          schema:
            type: integer
      responses:
        "200":
          description: Lista de profesores obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      description: El identificador único del profesor.
                    nombre:
                      type: string
                      description: El nombre del profesor.
                    correo:
                      type: string
                      description: El correo electrónico del profesor.
        "403":
          description: Acceso denegado
          content:
            text/plain:
              schema:
                type: string
                example: Usuario no autenticado o sin permisos.
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example: Error al obtener los profesores.
...